// src/app/core/service-content-type.ts
export class ServiceContentType {
    json: string = "application/json";
    formUrlEncode: string = "application/x-www-form-urlencoded";
    atomXml: string = "application/atom+xml"
    ecmascript: string = "application/ecmascript"
    javascript: string = "application/javascript"
    octetStream: string = "application/octet-stream"
    ogg: string = "application/ogg"
    pdf: string = "application/pdf"
    postscript: string = "application/postscript"
    rdfXml: string = "application/rdf+xml"
    rssXml: string = "application/rss+xml"
    soapXml: string = "application/soap+xml"
    fontWoff: string = "application/font-woff"
    xYaml: string = "application/x-yaml"
    xhtmlXml: string = "application/xhtml+xml"
    xml: string = "application/xml"
    xmlDtd: string = "application/xml-dtd"
    xopXml: string = "application/xop+xml"
    zip: string = "application/zip"
    gzip: string = "application/gzip"
    graphql: string = "application/graphql"


}